Checkout page organized and bug fixed
