+++ Frontend/src/pages/checkout/Checkout.js
@@ -13,14 +13,14 @@ import Typography from '@mui/material/Typography';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
 import ChevronLeftRoundedIcon from '@mui/icons-material/ChevronLeftRounded';
 import ChevronRightRoundedIcon from '@mui/icons-material/ChevronRightRounded';
-import AddressForm from './components/AddressForm';
-import getCheckoutTheme from './theme/getCheckoutTheme';
-import Info from './components/Info';
-import InfoMobile from './components/InfoMobile';
-import PaymentForm from './components/PaymentForm';
-import Review from './components/Review';
-import SitemarkIcon from './components/SitemarkIcon';
+import AddressForm from '../../components/AddressForm';
+import getCheckoutTheme from '../../theme/getCheckoutTheme';
+import Info from '../../components/Info';
+import InfoMobile from '../../components/InfoMobile';
+import PaymentForm from '../../components/PaymentForm';
+import Review from '../../components/Review';
 import TemplateFrame from './TemplateFrame';
+import { LogoIcon } from '../sign/CustomIcons';
 
 const steps = ['Shipping address', 'Payment details', 'Review your order'];
 function getStepContent(step) {
@@ -94,7 +94,7 @@ export default function Checkout() {
               gap: 4,
             }}
           >
-            <SitemarkIcon />
+            <LogoIcon />
             <Box
               sx={{
                 display: 'flex',
