+++ Frontend/src/theme/customizations/feedback.js
@@ -1,5 +1,5 @@
 import { alpha } from '@mui/material/styles';
-import { gray, orange } from '../themePrimitives';
+import { gray, orange } from "../themePrimitives";
 
 /* eslint-disable import/prefer-default-export */
 export const feedbackCustomizations = {
@@ -8,7 +8,7 @@ export const feedbackCustomizations = {
       root: ({ theme }) => ({
         borderRadius: 10,
         backgroundColor: orange[100],
-        color: (theme.vars || theme).palette.text.primary,
+        color: theme.palette.text.primary,
         border: `1px solid ${alpha(orange[300], 0.5)}`,
         '& .MuiAlert-icon': {
           color: orange[500],
@@ -26,7 +26,7 @@ export const feedbackCustomizations = {
         '& .MuiDialog-paper': {
           borderRadius: '10px',
           border: '1px solid',
-          borderColor: (theme.vars || theme).palette.divider,
+          borderColor: theme.palette.divider,
         },
       }),
     },
