+++ Frontend/src/theme/customizations/navigation.js
@@ -7,14 +7,14 @@ import { menuItemClasses } from '@mui/material/MenuItem';
 import { selectClasses } from '@mui/material/Select';
 import { tabClasses } from '@mui/material/Tab';
 import UnfoldMoreRoundedIcon from '@mui/icons-material/UnfoldMoreRounded';
-import { gray, brand } from '../themePrimitives';
+import { gray, brand } from "../themePrimitives";
 
 /* eslint-disable import/prefer-default-export */
 export const navigationCustomizations = {
   MuiMenuItem: {
     styleOverrides: {
       root: ({ theme }) => ({
-        borderRadius: (theme.vars || theme).shape.borderRadius,
+        borderRadius: theme.shape.borderRadius,
         padding: '6px 8px',
         [`&.${menuItemClasses.focusVisible}`]: {
           backgroundColor: 'transparent',
@@ -37,8 +37,8 @@ export const navigationCustomizations = {
       },
       paper: ({ theme }) => ({
         marginTop: '4px',
-        borderRadius: (theme.vars || theme).shape.borderRadius,
-        border: `1px solid ${(theme.vars || theme).palette.divider}`,
+        borderRadius: theme.shape.borderRadius,
+        border: `1px solid ${theme.palette.divider}`,
         backgroundImage: 'none',
         background: 'hsl(0, 0%, 100%)',
         boxShadow:
@@ -64,14 +64,14 @@ export const navigationCustomizations = {
     },
     styleOverrides: {
       root: ({ theme }) => ({
-        borderRadius: (theme.vars || theme).shape.borderRadius,
+        borderRadius: theme.shape.borderRadius,
         border: '1px solid',
         borderColor: gray[200],
-        backgroundColor: (theme.vars || theme).palette.background.paper,
+        backgroundColor: theme.palette.background.paper,
         boxShadow: `inset 0 1px 0 1px hsla(220, 0%, 100%, 0.6), inset 0 -1px 0 1px hsla(220, 35%, 90%, 0.5)`,
         '&:hover': {
           borderColor: gray[300],
-          backgroundColor: (theme.vars || theme).palette.background.paper,
+          backgroundColor: theme.palette.background.paper,
           boxShadow: 'none',
         },
         [`&.${selectClasses.focused}`]: {
@@ -82,13 +82,13 @@ export const navigationCustomizations = {
           display: 'none',
         },
         ...theme.applyStyles('dark', {
-          borderRadius: (theme.vars || theme).shape.borderRadius,
+          borderRadius: theme.shape.borderRadius,
           borderColor: gray[700],
-          backgroundColor: (theme.vars || theme).palette.background.paper,
+          backgroundColor: theme.palette.background.paper,
           boxShadow: `inset 0 1px 0 1px ${alpha(gray[700], 0.15)}, inset 0 -1px 0 1px hsla(220, 0%, 0%, 0.7)`,
           '&:hover': {
             borderColor: alpha(gray[700], 0.7),
-            backgroundColor: (theme.vars || theme).palette.background.paper,
+            backgroundColor: theme.palette.background.paper,
             boxShadow: 'none',
           },
           [`&.${selectClasses.focused}`]: {
@@ -119,7 +119,7 @@ export const navigationCustomizations = {
     },
     styleOverrides: {
       root: ({ theme }) => ({
-        color: (theme.vars || theme).palette.text.primary,
+        color: theme.palette.text.primary,
         fontWeight: 500,
         position: 'relative',
         textDecoration: 'none',
@@ -131,7 +131,7 @@ export const navigationCustomizations = {
           height: '1px',
           bottom: 0,
           left: 0,
-          backgroundColor: (theme.vars || theme).palette.text.secondary,
+          backgroundColor: theme.palette.text.secondary,
           opacity: 0.3,
           transition: 'width 0.3s ease, opacity 0.3s ease',
         },
@@ -149,7 +149,7 @@ export const navigationCustomizations = {
   MuiDrawer: {
     styleOverrides: {
       paper: ({ theme }) => ({
-        backgroundColor: (theme.vars || theme).palette.background.default,
+        backgroundColor: theme.palette.background.default,
       }),
     },
   },
@@ -158,12 +158,12 @@ export const navigationCustomizations = {
       root: ({ theme }) => ({
         '&.Mui-selected': {
           color: 'white',
-          backgroundColor: (theme.vars || theme).palette.grey[900],
+          backgroundColor: theme.palette.grey[900],
         },
         ...theme.applyStyles('dark', {
           '&.Mui-selected': {
             color: 'black',
-            backgroundColor: (theme.vars || theme).palette.grey[50],
+            backgroundColor: theme.palette.grey[50],
           },
         }),
       }),
@@ -173,9 +173,9 @@ export const navigationCustomizations = {
     styleOverrides: {
       root: { minHeight: 'fit-content' },
       indicator: ({ theme }) => ({
-        backgroundColor: (theme.vars || theme).palette.grey[800],
+        backgroundColor: theme.palette.grey[800],
         ...theme.applyStyles('dark', {
-          backgroundColor: (theme.vars || theme).palette.grey[200],
+          backgroundColor: theme.palette.grey[200],
         }),
       }),
     },
@@ -188,12 +188,12 @@ export const navigationCustomizations = {
         textTransform: 'none',
         minWidth: 'fit-content',
         minHeight: 'fit-content',
-        color: (theme.vars || theme).palette.text.secondary,
-        borderRadius: (theme.vars || theme).shape.borderRadius,
+        color: theme.palette.text.secondary,
+        borderRadius: theme.shape.borderRadius,
         border: '1px solid',
         borderColor: 'transparent',
         ':hover': {
-          color: (theme.vars || theme).palette.text.primary,
+          color: theme.palette.text.primary,
           backgroundColor: gray[100],
           borderColor: gray[200],
         },
@@ -202,7 +202,7 @@ export const navigationCustomizations = {
         },
         ...theme.applyStyles('dark', {
           ':hover': {
-            color: (theme.vars || theme).palette.text.primary,
+            color: theme.palette.text.primary,
             backgroundColor: gray[800],
             borderColor: gray[700],
           },
@@ -217,7 +217,7 @@ export const navigationCustomizations = {
     styleOverrides: {
       line: ({ theme }) => ({
         borderTop: '1px solid',
-        borderColor: (theme.vars || theme).palette.divider,
+        borderColor: theme.palette.divider,
         flex: 1,
         borderRadius: '99px',
       }),
@@ -236,21 +236,21 @@ export const navigationCustomizations = {
         },
         '&.Mui-active': {
           border: 'none',
-          color: (theme.vars || theme).palette.primary.main,
+          color: theme.palette.primary.main,
         },
         '&.Mui-completed': {
           border: 'none',
-          color: (theme.vars || theme).palette.success.main,
+          color: theme.palette.success.main,
         },
         ...theme.applyStyles('dark', {
           border: `1px solid ${gray[700]}`,
           '&.Mui-active': {
             border: 'none',
-            color: (theme.vars || theme).palette.primary.light,
+            color: theme.palette.primary.light,
           },
           '&.Mui-completed': {
             border: 'none',
-            color: (theme.vars || theme).palette.success.light,
+            color: theme.palette.success.light,
           },
         }),
         variants: [
